PMM-720: encapsulate stats, separate them from profiler
